# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: NEMSalon API
  version: 1.0.0
  description: Contract-first API for web and mobile clients.
servers:
  - url: https://api.nemsalon.dk
    description: Production
  - url: https://staging-api.nemsalon.dk
    description: Staging
  - url: http://localhost:3000/api
    description: Local
...
paths:
...
  /v1/salons/{salonId}:
    get:
      tags: [Salons]
      summary: Get salon
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SalonId'
      responses:
        '200':
          description: Salon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Salon'
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Salons]
      summary: Update salon
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SalonId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalonUpdate'
      responses:
        '200':
          description: Updated salon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Salon'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/salons/{salonId}/business-hours:
    get:
      tags: [Salons]
      summary: Get salon business hours
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SalonId'
      responses:
        '200':
          description: Business hours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessHoursWeekly'
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    put:
      tags: [Salons]
      summary: Replace salon business hours
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/SalonId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessHoursWeekly'
      responses:
        '200':
          description: Updated business hours
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessHoursWeekly'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/services:
    get:
      tags: [Services]
      summary: List services
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Service list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceList'
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Services]
      summary: Create service
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreate'
      responses:
        '201':
          description: Created service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/services/{serviceId}:
    get:
      tags: [Services]
      summary: Get service
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      responses:
        '200':
          description: Service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Services]
      summary: Update service
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdate'
      responses:
        '200':
          description: Updated service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/staff:
    get:
      tags: [Staff]
      summary: List staff
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Staff list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffList'
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Staff]
      summary: Create staff
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffCreate'
      responses:
        '201':
          description: Created staff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Staff'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/staff/{staffId}/services:
    get:
      tags: [Staff]
      summary: List services assigned to staff
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/StaffId'
      responses:
        '200':
          description: Staff services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffServicesResponse'
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Staff]
      summary: Assign services to staff
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/StaffId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffServicesAssignRequest'
      responses:
        '200':
          description: Assigned services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffServicesResponse'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/customers:
    get:
      tags: [Customers]
      summary: List customers
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Customer list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Customers]
      summary: Create customer
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/customers/{customerId}:
    get:
      tags: [Customers]
      summary: Get customer
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Customers]
      summary: Update customer
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
...
  /v1/bookings:
    get:
      tags: [Bookings]
      summary: List bookings
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Booking list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingList'
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Bookings]
      summary: Create booking
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreate'
      responses:
        '201':
          description: Created booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
  /v1/bookings/{bookingId}:
    get:
      tags: [Bookings]
      summary: Get booking
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Bookings]
      summary: Update booking
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdate'
      responses:
        '200':
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
...
  /v1/bookings/{bookingId}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel a booking
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingRequest'
      responses:
        '200':
          description: Cancelled booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /v1/bookings/{bookingId}/reschedule:
    post:
      tags: [Bookings]
      summary: Reschedule a booking
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RescheduleBookingRequest'
      responses:
        '200':
          description: Rescheduled booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    SalonId:
      name: salonId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    StaffId:
      name: staffId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CustomerId:
      name: customerId
      in: path
      required: true
      schema: { type: string, format: uuid }
    BookingId:
      name: bookingId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
...
    Salon:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [draft, active] }
        currency: { type: string, minLength: 3, maxLength: 3, pattern: '^[A-Z]{3}$' }
        locale: { type: string }
        timezone: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    SalonUpdate:
      type: object
      properties:
        name: { type: string }
        currency: { type: string, minLength: 3, maxLength: 3, pattern: '^[A-Z]{3}$' }
        locale: { type: string }
        timezone: { type: string }
    BusinessHoursDay:
      type: object
      required: [day, startTime, endTime, enabled]
      properties:
        day: { type: string, enum: [mon, tue, wed, thu, fri, sat, sun] }
        startTime: { type: string, example: '09:00' }
        endTime: { type: string, example: '17:00' }
        enabled: { type: boolean }
    BusinessHoursWeekly:
      type: object
      required: [weekly]
      properties:
        weekly:
          type: array
          items: { $ref: '#/components/schemas/BusinessHoursDay' }
...
    Service:
      type: object
      required: [id, salonId, name, price, durationMinutes, currency]
      properties:
        id: { type: string, format: uuid }
        salonId: { type: string, format: uuid }
        name: { type: string }
        price: { type: integer, description: 'Minor units (e.g., øre)' }
        currency: { type: string, minLength: 3, maxLength: 3, pattern: '^[A-Z]{3}$' }
        durationMinutes: { type: integer }
        bufferMinutes: { type: integer }
        active: { type: boolean }
    ServiceCreate:
      type: object
      required: [name, price, durationMinutes, currency]
      properties:
        name: { type: string }
        price: { type: integer, description: 'Minor units (e.g., øre)' }
        currency: { type: string, minLength: 3, maxLength: 3, pattern: '^[A-Z]{3}$' }
        durationMinutes: { type: integer }
        bufferMinutes: { type: integer }
        active: { type: boolean }
    ServiceUpdate:
      type: object
      properties:
        name: { type: string }
        price: { type: integer, description: 'Minor units (e.g., øre)' }
        currency: { type: string, minLength: 3, maxLength: 3, pattern: '^[A-Z]{3}$' }
        durationMinutes: { type: integer }
        bufferMinutes: { type: integer }
        active: { type: boolean }
    ServiceList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Service' }
    Staff:
      type: object
      required: [id, salonId, name, role]
      properties:
        id: { type: string, format: uuid }
        salonId: { type: string, format: uuid }
        name: { type: string }
        role: { type: string, enum: [owner, admin, staff] }
        active: { type: boolean }
    StaffCreate:
      type: object
      required: [name, role]
      properties:
        name: { type: string }
        role: { type: string, enum: [owner, admin, staff] }
        active: { type: boolean }
    StaffServicesAssignRequest:
      type: object
      required: [serviceIds]
      properties:
        serviceIds:
          type: array
          items: { type: string, format: uuid }
    StaffServicesResponse:
      type: object
      required: [staffId, serviceIds]
      properties:
        staffId: { type: string, format: uuid }
        serviceIds:
          type: array
          items: { type: string, format: uuid }
    StaffList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Staff' }
    Customer:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        notes: { type: string }
    CustomerCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        notes: { type: string }
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        notes: { type: string }
    CustomerList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
...
    Booking:
      type: object
      required: [id, salonId, customerId, staffId, serviceId, startTime, endTime, status, totalAmount, currency]
      properties:
        id: { type: string, format: uuid }
        salonId: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        staffId: { type: string, format: uuid }
        serviceId: { type: string, format: uuid }
        startTime: { type: string, format: date-time, description: 'UTC ISO timestamp' }
        endTime: { type: string, format: date-time, description: 'UTC ISO timestamp' }
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        totalAmount: { type: integer, description: 'Minor units (e.g., øre)' }
        currency: { type: string, minLength: 3, maxLength: 3, pattern: '^[A-Z]{3}$' }
        notes: { type: string }
    BookingCreate:
      type: object
      required: [staffId, serviceId, startUtc]
      properties:
        customerId: { type: string, format: uuid }
        customer: { $ref: '#/components/schemas/CustomerCreate' }
        staffId: { type: string, format: uuid }
        serviceId: { type: string, format: uuid }
        startUtc:
          type: string
          format: date-time
          description: 'UTC ISO timestamp, must align to 15-minute grid.'
        endUtc:
          type: string
          format: date-time
          description: Optional; server computes endTime based on service duration.
        startTime:
          type: string
          format: date-time
          description: Deprecated alias for startUtc.
        endTime:
          type: string
          format: date-time
          description: Deprecated alias for endUtc.
        notes: { type: string }
      oneOf:
        - required: [customerId]
        - required: [customer]
    BookingResponse:
      type: object
      required: [booking]
      properties:
        booking:
          $ref: '#/components/schemas/Booking'
    CancelBookingRequest:
      type: object
      properties:
        reasonKey: { type: string }
        note: { type: string }
    RescheduleBookingRequest:
      type: object
      required: [staffId, startUtc]
      properties:
        staffId: { type: string, format: uuid }
        startUtc:
          type: string
          format: date-time
          description: 'UTC ISO timestamp, must align to 15-minute grid.'
    BookingUpdate:
      type: object
      properties:
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        notes: { type: string }
    BookingList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Booking' }
...